requirements.txt
#Dependencias
torch
torch-geometric
networkx
numpy
pandas
openpyxl
scikit-learn
PyYAML